Update for customizable RBAC
Roles are now customizable in the backend, and this codebase handles that. A user still has a single role, but now we give a list of all roles that role "inherits" from.

The old attribute userRole is renamed userAssignedRole. It's now a straight string instead of an enum.

There are two new attributes to look at:

- userRoles is a list of all roles that this user "inherits". Instead of checking roles with < and >, you should check for the existence of the role in this list.
- userPermissions is a list of all permissions this user has. These are the four hardcoded permissions, e.g. SAML, deleting users, etc.